Backend/requirement.txt
fastapi==0.111.0                # Web framework for backend API
uvicorn==0.23.2                 # ASGI server to run FastAPI
google-cloud-speech==2.26.0     # Google Speech-to-Text
google-cloud-texttospeech==2.16.3 # Google Text-to-Speech
google-generativeai==0.7.2      # Google Gemini API
python-multipart==0.0.7         # For handling file uploads in FastAPI
pydub